Move icon options into own lib file
