+++ src/components/DynamicIcon.tsx
@@ -1,76 +1,4 @@
-import {
-    Utensils,
-    Film,
-    Tv,
-    Book,
-    Music,
-    MapPin,
-    Star,
-    Gamepad2,
-    Footprints,
-    Mountain,
-    Coffee,
-    Pizza,
-    Beer,
-    Wine,
-    Laptop,
-    Smartphone,
-    Code,
-    Palette,
-    Dumbbell,
-    Trophy,
-    Home,
-    Car,
-    ShoppingBag,
-    Gift,
-    Briefcase,
-    GraduationCap,
-    Plane,
-    Smile,
-    Heart,
-    Zap,
-} from 'lucide-react';
-import type { LucideIcon } from 'lucide-react';
-
-// Map of allowed icons
-const iconMap: Record<string, LucideIcon> = {
-    restaurant: Utensils,
-    movie: Film,
-    show: Tv,
-    book: Book,
-    album: Music,
-    location: MapPin,
-    hiking: Mountain,
-    activity: Footprints,
-    game: Gamepad2,
-    food: Utensils,
-    coffee: Coffee,
-    pizza: Pizza,
-    drink: Beer,
-    wine: Wine,
-    tech: Laptop,
-    mobile: Smartphone,
-    code: Code,
-    art: Palette,
-    gym: Dumbbell,
-    sports: Trophy,
-    home: Home,
-    car: Car,
-    shopping: ShoppingBag,
-    gift: Gift,
-    work: Briefcase,
-    school: GraduationCap,
-    travel: Plane,
-    fun: Smile,
-    love: Heart,
-    energy: Zap,
-    default: Star,
-};
-
-// Source of truth for selectable icons. Filter default so it doesn't show up as a choice
-export const ICON_OPTIONS = Object.keys(iconMap)
-    .filter((key) => key !== 'default')
-    .sort();
+import { iconMap } from '@/lib/icons';
 
 type Props = {
     name: string;
